[Refactor] 친구목록 조회 시 userId대신 loginId반환하도록 변경#34
