+++ src/test/java/me/snaptime/snap/repository/SnapPagingRepositoryImplTest.java
@@ -21,6 +21,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.TestPropertySource;
@@ -37,6 +38,7 @@ import static org.springframework.test.util.AssertionErrors.fail;
 @DataJpaTest
 @Import({QueryDslConfig.class, JpaAuditingConfig.class})
 @TestPropertySource(locations = "classpath:application-test.yml")
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 public class SnapPagingRepositoryImplTest {
 
     @Autowired
