+++ src/test/java/me/snaptime/social/controller/FriendShipControllerTest.java
@@ -1,4 +1,4 @@
-package me.snaptime.social.controllerTest;
+package me.snaptime.social.controller;
 
 import com.google.gson.Gson;
 import me.snaptime.common.config.SecurityConfig;
@@ -6,7 +6,6 @@ import me.snaptime.common.exception.customs.CustomException;
 import me.snaptime.common.exception.customs.ExceptionCode;
 import me.snaptime.common.jwt.JwtProvider;
 import me.snaptime.social.common.FriendSearchType;
-import me.snaptime.social.controller.FriendShipController;
 import me.snaptime.social.data.dto.req.AcceptFollowReqDto;
 import me.snaptime.social.service.FriendShipService;
 import me.snaptime.user.service.UserDetailsServiceImpl;
